\name{GasFlow}
\alias{GasFlow}
\docType{data}
\title{
Flows in pipelines
}
\description{
Daily flows of natural gas in three pipelines
}
\usage{data(GasFlow)}
\format{
  A data frame with 342 observations on the following 3 variables.
  \describe{
    \item{\code{Flow1}}{a numeric vector}
    \item{\code{Flow2}}{a numeric vector}
    \item{\code{Flow3}}{a numeric vector}
  }
}
\details{
Dataset used in:\cr
Example 4.3\cr
Example 5.9\cr
Example 5.10\cr
Example 8.2\cr
}
\source{
David Ruppert's textbook web page:\cr
\url{http://people.orie.cornell.edu/~davidr/SDAFE/data/index.html}
}
\references{
Statistics and Data Analysis for Financial Engineering by David Ruppert, Springer 2010
}
\examples{
data(GasFlow)
head(GasFlow)
}
\keyword{datasets}
