\name{sdafe-package}
\alias{sdafe-package}
\alias{sdafe}
\docType{package}
\title{Statistics and Data Analysis for Financial Engineering}
\description{
The sdafe package is an R companion to Statistics and 
	Data Analysis for Financial Engineering by David Ruppert (Springer, 2010).
	The package includes script files, functions, and data sets to reproduce 
	most examples, figures, and tables from the text.
}
\details{
\tabular{ll}{
Package: \tab sdafe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-23\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Guy Yollin\cr
Maintainer: Guy Yollin <gyollin@r-programming.org>
}
\references{
Statistics and Data Analysis for Financial Engineering, David Ruppert, Springer, 2010
}
\keyword{ package }
\seealso{
David Ruppert's textbook web page:\cr
\url{http://people.orie.cornell.edu/~davidr/SDAFE/data/index.html}
}
\examples{
# display figure
Figure_2.1()
}
