\name{sdafe-package}
\alias{sdafe-package}
\alias{sdafe}
\docType{package}
\title{Statistics and Data Analysis for Financial Engineering}
\description{
The SDAFE package is an R companion to Statistics and 
	Data Analysis for Financial Engineering by David Ruppert (2011, Springer).
	The package includes script files, functions, and data sets to reproduce 
	most examples, figures, and tables from the text.
}
\details{
\tabular{ll}{
Package: \tab sdafe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-23\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
